Tests: Updated code to support includes & dependencies on the library down to the testing executable; Restructure tests to allow global server instance
